refactor: fix clippy warnings and complete wrapper handler extraction
- Fix clippy warnings for too many arguments by introducing WrapperContext struct
- Move remaining wrapper import logic from Bundler to wrapper handler
- Extract handle_symbol_imports_from_multiple -> handle_symbol_imports_from_wrapper (560+ lines)
- Remove unused should_expose_on_namespace method from Bundler
- Centralize all wrapper import transformation logic in handlers/wrapper.rs
- Maintain full test compatibility (132/132 tests passing)
